backend/django_service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-jose==3.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
